requirements.txt
pandas
pyarrow
psycopg[binary]
python-dotenv